+++ src/Plugins/SmartStore.OfflinePayment/Models/ConfigurationModel.cs
@@ -3,11 +3,14 @@ using System.Web.Mvc;
 using SmartStore.OfflinePayment.Settings;
 using SmartStore.Web.Framework;
 using SmartStore.Web.Framework.Modelling;
+using System.ComponentModel.DataAnnotations;
 
 namespace SmartStore.OfflinePayment.Models
 { 
 	public abstract class ConfigurationModelBase : ModelBase
     {
+		public string PrimaryStoreCurrencyCode { get; set; }
+
 		[AllowHtml]
 		[SmartResourceDisplayName("Plugins.SmartStore.OfflinePayment.DescriptionText")]
 		public string DescriptionText { get; set; }
@@ -17,7 +20,11 @@ namespace SmartStore.OfflinePayment.Models
 
 		[SmartResourceDisplayName("Plugins.SmartStore.OfflinePayment.AdditionalFeePercentage")]
 		public bool AdditionalFeePercentage { get; set; }
-    }
+
+		[SmartResourceDisplayName("Plugins.SmartStore.OfflinePayment.PaymentMethodLogo")]
+		[UIHint("Picture")]
+		public int PaymentMethodLogo { get; set; }
+	}
 
 	public class CashOnDeliveryConfigurationModel : ConfigurationModelBase
 	{