+++ src/Presentation/SmartStore.Web/Controllers/CommonController.cs
@@ -406,8 +406,9 @@ namespace SmartStore.Web.Controllers
 				DisplayAdminLink = _services.Permissions.Authorize(StandardPermissionProvider.AccessAdminPanel),
 				ShoppingCartEnabled = _services.Permissions.Authorize(StandardPermissionProvider.EnableShoppingCart) && _shoppingCartSettings.MiniShoppingCartEnabled,
 				WishlistEnabled = _services.Permissions.Authorize(StandardPermissionProvider.EnableWishlist),
-                CompareProductsEnabled = _catalogSettings.CompareProductsEnabled            
-            };
+                CompareProductsEnabled = _catalogSettings.CompareProductsEnabled,
+				PublicStoreNavigationAllowed = _services.Permissions.Authorize(StandardPermissionProvider.PublicStoreAllowNavigation)
+			};
 
 			return PartialView(model);
         }
@@ -637,8 +638,6 @@ namespace SmartStore.Web.Controllers
             var disallowPaths = new List<string>()
             {
                 "/bin/",
-                "/Content/files/",
-                "/Content/files/ExportImport/",
 				"/Exchange/",
                 "/Country/GetStatesByCountryId",
                 "/Install",